moving io.tools in api.tools enforcing no ImportError unless ran from the same file (script mode)
