+++ radis/api/geisaapi.py
@@ -24,8 +24,11 @@ try:
     from .cache_files import load_h5_cache_file, save_to_hdf
     from .tools import drop_object_format_columns, parse_hitran_file
 except ImportError:
-    from radis.api.cache_files import load_h5_cache_file, save_to_hdf
-    from radis.io.tools import drop_object_format_columns, parse_hitran_file
+    if __name__ == "__main__":  # running from this file, as a script
+        from radis.api.cache_files import load_h5_cache_file, save_to_hdf
+        from radis.io.tools import drop_object_format_columns, parse_hitran_file
+    else:
+        raise
 # from typing import Union
 from radis.api.dbmanager import DatabaseManager
 from radis.api.hdf5 import DataFileManager
