+++ radis/api/hitempapi.py
@@ -27,18 +27,21 @@ try:
         replace_PQR_with_m101,
     )
 except ImportError:  # ran from here
-    from radis.api.hitranapi import (
-        columns_2004,
-        parse_global_quanta,
-        parse_local_quanta,
-    )
-    from radis.io.tools import (
-        _create_dtype,
-        _get_linereturnformat,
-        _ndarray2df,
-        replace_PQR_with_m101,
-    )
-    from radis.api.dbmanager import DatabaseManager
+    if __name__ == "__main__":  # running from this file, as a script
+        from radis.api.dbmanager import DatabaseManager
+        from radis.api.hitranapi import (
+            columns_2004,
+            parse_global_quanta,
+            parse_local_quanta,
+        )
+        from radis.io.tools import (
+            _create_dtype,
+            _get_linereturnformat,
+            _ndarray2df,
+            replace_PQR_with_m101,
+        )
+    else:
+        raise
 
 from radis.db import MOLECULES_LIST_NONEQUILIBRIUM
 from radis.misc.progress_bar import ProgressBar
