+++ radis/api/hitranapi.py
@@ -31,9 +31,6 @@ from os.path import abspath, exists, getmtime, join, split
 import pandas as pd
 from numpy import int64
 
-from radis.api.cache_files import load_h5_cache_file, save_to_hdf
-from radis.api.dbmanager import DatabaseManager
-from radis.api.hdf5 import DataFileManager
 from radis.db.classes import (  # get_molecule_identifier,
     HITRAN_CLASS1,
     HITRAN_CLASS2,
@@ -53,11 +50,28 @@ from radis.db.classes import (  # get_molecule_identifier,
     HITRAN_GROUP6,
     get_molecule,
 )
-from radis.io.tools import (
-    drop_object_format_columns,
-    parse_hitran_file,
-    replace_PQR_with_m101,
-)
+
+try:
+    from .cache_files import load_h5_cache_file, save_to_hdf
+    from .dbmanager import DatabaseManager
+    from .hdf5 import DataFileManager
+    from .tools import (
+        drop_object_format_columns,
+        parse_hitran_file,
+        replace_PQR_with_m101,
+    )
+except ImportError:
+    if __name__ == "__main__":  # running from this file, as a script
+        from radis.api.cache_files import load_h5_cache_file, save_to_hdf
+        from radis.api.dbmanager import DatabaseManager
+        from radis.api.hdf5 import DataFileManager
+        from radis.api.tools import (
+            drop_object_format_columns,
+            parse_hitran_file,
+            replace_PQR_with_m101,
+        )
+    else:
+        raise
 
 # %% Parsing functions
